// management-frontend/api/property/models/User.ts
/* tslint:disable */
/* eslint-disable */
/**
 * GM Properties API
 * OpenAPI-Spezifikation der GM Properties-API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface User
 */
export interface User {
    /**
     * Id
     * @type {string}
     * @memberof User
     */
    id?: string;
    /**
     * Mail Adresse des Nutzers
     * @type {string}
     * @memberof User
     */
    mail?: string;
    /**
     * Name des Nutzers
     * @type {string}
     * @memberof User
     */
    name?: string;
    /**
     * Role des Nutzers
     * @type {string}
     * @memberof User
     */
    role?: UserRoleEnum;
}


/**
 * @export
 */
export const UserRoleEnum = {
    Admin: 'Admin',
    Facility: 'Facility',
    Finance: 'Finance',
    Operations: 'Operations'
} as const;
export type UserRoleEnum = typeof UserRoleEnum[keyof typeof UserRoleEnum];


/**
 * Check if a given object implements the User interface.
 */
export function instanceOfUser(value: object): value is User {
    return true;
}

export function UserFromJSON(json: any): User {
    return UserFromJSONTyped(json, false);
}

export function UserFromJSONTyped(json: any, ignoreDiscriminator: boolean): User {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'mail': json['mail'] == null ? undefined : json['mail'],
        'name': json['name'] == null ? undefined : json['name'],
        'role': json['role'] == null ? undefined : json['role'],
    };
}

  export function UserToJSON(json: any): User {
      return UserToJSONTyped(json, false);
  }

  export function UserToJSONTyped(value?: User | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'mail': value['mail'],
        'name': value['name'],
        'role': value['role'],
    };
}